propagated changes of Sunday Jun 05 2005 from the SQL_2-8 release branch to the development trunk

+++ sql/src/odbc/samples/odbcsample1.c
@@ -56,7 +56,7 @@ prerr(SQLSMALLINT tpe, SQLHANDLE hnd, const char *func, const char *pref)
 		if (msglen >= (signed int) sizeof(msg))
 			fprintf(stderr, "(message truncated)\n");
 	case SQL_SUCCESS:
-		fprintf(stderr, "%s: %s: SQLstate %s, Errnr %d, Message %s\n", func, pref, state, errnr, msg);
+		fprintf(stderr, "%s: %s: SQLstate %s, Errnr %d, Message %s\n", func, pref, state, (int)errnr, msg);
 		break;
 	case SQL_INVALID_HANDLE:
 		fprintf(stderr, "%s: %s, invalid handle passed to error function\n", func, pref);
